[TASK] Migrate part of `CommentTests` to unit tests
Part of #758.
