+++ tests/Comment/CommentTest.php
@@ -5,9 +5,7 @@ declare(strict_types=1);
 namespace Sabberworm\CSS\Tests\Comment;
 
 use PHPUnit\Framework\TestCase;
-use Sabberworm\CSS\Comment\Comment;
 use Sabberworm\CSS\OutputFormat;
-use Sabberworm\CSS\Renderable;
 use Sabberworm\CSS\Tests\ParserTest as TestsParserTest;
 
 /**
@@ -17,97 +15,6 @@ use Sabberworm\CSS\Tests\ParserTest as TestsParserTest;
  */
 final class CommentTest extends TestCase
 {
-    /**
-     * @test
-     */
-    public function implementsRenderable(): void
-    {
-        $subject = new Comment();
-
-        self::assertInstanceOf(Renderable::class, $subject);
-    }
-
-    /**
-     * @test
-     */
-    public function getCommentOnEmptyInstanceReturnsReturnsEmptyString(): void
-    {
-        $subject = new Comment();
-
-        self::assertSame('', $subject->getComment());
-    }
-
-    /**
-     * @test
-     */
-    public function getCommentInitiallyReturnsCommentPassedToConstructor(): void
-    {
-        $comment = 'There is no spoon.';
-        $subject = new Comment($comment);
-
-        self::assertSame($comment, $subject->getComment());
-    }
-
-    /**
-     * @test
-     */
-    public function setCommentSetsComments(): void
-    {
-        $comment = 'There is no spoon.';
-        $subject = new Comment();
-
-        $subject->setComment($comment);
-
-        self::assertSame($comment, $subject->getComment());
-    }
-
-    /**
-     * @test
-     */
-    public function getLineNoOnEmptyInstanceReturnsReturnsZero(): void
-    {
-        $subject = new Comment();
-
-        self::assertSame(0, $subject->getLineNo());
-    }
-
-    /**
-     * @test
-     */
-    public function getLineNoInitiallyReturnsLineNumberPassedToConstructor(): void
-    {
-        $lineNumber = 42;
-        $subject = new Comment('', $lineNumber);
-
-        self::assertSame($lineNumber, $subject->getLineNo());
-    }
-
-    /**
-     * @test
-     */
-    public function toStringRendersCommentEnclosedInCommentDelimiters(): void
-    {
-        $comment = 'There is no spoon.';
-        $subject = new Comment();
-
-        $subject->setComment($comment);
-
-        self::assertSame('/*' . $comment . '*/', (string) $subject);
-    }
-
-    /**
-     * @test
-     */
-    public function renderRendersCommentEnclosedInCommentDelimiters(): void
-    {
-        $comment = 'There is no spoon.';
-        $subject = new Comment();
-
-        $subject->setComment($comment);
-
-        self::assertSame('/*' . $comment . '*/', $subject->render(new OutputFormat()));
-    }
-
     /**
      * @test
      */
